refactor(config/appdir): gather the Prefix type into its own files
The Prefix type's material was scattered: PrefixUser sat in
appdir.go, the FHS PrefixLocal, PrefixSystem and PrefixOptional in
appdir_fhs.go, while its tests were split between prefix_test.go and
appdir_test.go. Gather it so the type has a clear home.

The constants move to per-platform files. The user-mode sentinel
PrefixUser and the FHS trio go to const_unix.go, behind the same
!windows tag appdir_fhs.go already carried, joined by a new
isWellKnown method folding the well-known match Validate used to
spell out inline; Validate now delegates to it. PrefixUser is a mode
selector, not a shared root: unix spells it "~", and the Windows
spelling lands with const_windows.go, so it lives beside the FHS
trio rather than in a portable file. Splitting the constants and
their membership check per platform lets each define its own set.

The portable Prefix tests follow. TestNewPrefix and TestSysPrefix —
which assert Prefix values and validation errors, with no
platform-specific path composition — move from appdir_test.go to
prefix_test.go, alongside the TestPrefixValidate and
TestPrefixUserMode already there. The remaining prefix tests stay in
appdir_test.go: TestSetSysPrefix and TestSetSysPrefixUserMode assert
FHS and XDG paths, and TestNewPrefixAbsError depends on removing the
working directory, so all three are inherently unix-specific;
appdir_test.go takes a !windows tag to match.

No behaviour changes; this only relocates constants and tests so the
Prefix type is self-contained and ready to grow platform variants.
