+++ config/appdir/appdir_test.go
@@ -1,3 +1,5 @@
+//go:build !windows
+
 package appdir_test
 
 import (
@@ -18,7 +20,6 @@ var _ core.TestCase = userDirTestCase{}
 var _ core.TestCase = userDirErrTestCase{}
 var _ core.TestCase = sysDirTestCase{}
 var _ core.TestCase = sysUserModeTestCase{}
-var _ core.TestCase = newPrefixTestCase{}
 var _ core.TestCase = setSysPrefixTestCase{}
 
 // userDirTestCase tests the UserFooDir functions honouring their
@@ -388,79 +389,6 @@ func TestSysDirUserMode(t *testing.T) {
 	core.RunTestCases(t, testCases)
 }
 
-// newPrefixTestCase tests [appdir.NewPrefix] validation and
-// path resolution.
-type newPrefixTestCase struct {
-	wantErrIs error
-	dir       string
-	name      string
-	want      appdir.Prefix
-}
-
-func (tc newPrefixTestCase) Name() string {
-	return tc.name
-}
-
-func (tc newPrefixTestCase) Test(t *testing.T) {
-	t.Helper()
-
-	got, err := appdir.NewPrefix(tc.dir)
-	if tc.wantErrIs != nil {
-		core.AssertErrorIs(t, err, tc.wantErrIs, "new prefix")
-		return
-	}
-
-	core.AssertMustNoError(t, err, "new prefix")
-	core.AssertEqual(t, tc.want, got, "prefix")
-}
-
-// newNewPrefixTestCase declares a row expected to succeed, with
-// want holding the resulting Prefix value.
-func newNewPrefixTestCase(name, dir string,
-	want appdir.Prefix) newPrefixTestCase {
-	return newPrefixTestCase{
-		dir:  dir,
-		name: name,
-		want: want,
-	}
-}
-
-// newNewPrefixTestCaseErr declares a row expected to fail.
-func newNewPrefixTestCaseErr(name, dir string,
-	wantErrIs error) newPrefixTestCase {
-	return newPrefixTestCase{
-		wantErrIs: wantErrIs,
-		dir:       dir,
-		name:      name,
-	}
-}
-
-func newPrefixTestCases(tmp, file, cwd string) []newPrefixTestCase {
-	return core.S(
-		newNewPrefixTestCase("user mode", "~", appdir.PrefixUser),
-		newNewPrefixTestCase("existing dir", tmp,
-			appdir.Prefix(tmp)),
-		newNewPrefixTestCase("relative path", ".",
-			appdir.Prefix(cwd)),
-		newNewPrefixTestCaseErr("missing path",
-			filepath.Join(tmp, "missing"), fs.ErrNotExist),
-		newNewPrefixTestCaseErr("regular file", file,
-			syscall.ENOTDIR),
-	)
-}
-
-func TestNewPrefix(t *testing.T) {
-	tmp := t.TempDir()
-	file := filepath.Join(tmp, "file")
-	err := os.WriteFile(file, []byte("x"), 0o600)
-	core.AssertMustNoError(t, err, "write file")
-
-	cwd, err := os.Getwd()
-	core.AssertMustNoError(t, err, "getwd")
-
-	core.RunTestCases(t, newPrefixTestCases(tmp, file, cwd))
-}
-
 // TestNewPrefixAbsError pins [appdir.NewPrefix] propagating the
 // filepath.Abs failure resolving a relative argument when the
 // working directory no longer exists.
@@ -551,17 +479,6 @@ func TestSetSysPrefix(t *testing.T) {
 	core.AssertMustNoError(t, err, "getwd")
 
 	core.RunTestCases(t, setSysPrefixTestCases(tmp, file, cwd))
-}
-
-// TestSysPrefix pins the getter reflecting the current default
-// Prefix.
-func TestSysPrefix(t *testing.T) {
-	core.AssertEqual(t, appdir.PrefixUser, appdir.SysPrefix(),
-		"default")
-
-	t.Cleanup(appdir.StubSysPrefix(appdir.PrefixSystem))
-	core.AssertEqual(t, appdir.PrefixSystem, appdir.SysPrefix(),
-		"stubbed")
 }
 
 // TestSetSysPrefixUserMode pins "~" returning the package to user
